Removed PRs from summaries (#14)
From the documentation:
> GitHub's REST API v3 considers every pull request an issue, but not
every issue is a pull request. For this reason, "Issues" endpoints may
return both issues and pull requests in the response. You can identify
pull requests by the pull_request key. Be aware that the id of a pull
request returned from "Issues" endpoints will be an issue id.

This commit filters objects that *have* a `pull_request` object